Make makefile a bit more clear;  Fix warnings with clang.

+++ emulators/c/linux-emulator-example.c
@@ -61,7 +61,7 @@ int main(void){
 
 	while(!is_halted(vm)){
 		if(vm_getc(vm, &output)){
-			putchar(output);
+			putchar((int)output);
 			fflush(stdout);
 		}
